+++ .gitignore
@@ -6,3 +6,4 @@ dist/*
 .github/workflows/publish-to-test-pipy-backup.yml
 **/*.lock
 *.tar.gz
+log/
