+++ activexendpoint/controllers/__init__.py
@@ -1,0 +1,3 @@
+from .putmetadata import put_metadata
+from .addcode import add_code
+from .methodexecution import method_exeution
