--- log/client-0.2024-06-19_10-00
@@ -1,14 +1,0 @@
-{
-    "timestamp": "2024-06-19 10:00:32,447",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "2lznxt554duyextli051a3yp3qkaz9ge",
-    "key": "6cbg48jxxgo7gfxf",
-    "size": 19492,
-    "response_time": 0.024288415908813477,
-    "metadata_service_time": 0.00888371467590332,
-    "peer_id": "mictlanx-peer-1"
-}
-
