--- log/client-0.2024-06-19_15-35
@@ -1,168 +1,0 @@
-{
-    "timestamp": "2024-06-19 14:34:47,155",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "2lznxt554duyextli051a3yp3qkaz9ge",
-    "key": "6cbg48jxxgo7gfxf",
-    "size": 19492,
-    "response_time": 0.0369873046875,
-    "metadata_service_time": 0.0199737548828125,
-    "peer_id": "mictlanx-peer-0"
-}
-
-{
-    "timestamp": "2024-06-19 14:36:01,149",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "2lznxt554duyextli051a3yp3qkaz9ge",
-    "key": "6cbg48jxxgo7gfxf",
-    "size": 19492,
-    "response_time": 0.019071340560913086,
-    "metadata_service_time": 0.008666276931762695,
-    "peer_id": "mictlanx-peer-0"
-}
-
-{
-    "timestamp": "2024-06-19 14:37:09,801",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "5p5r8u53utgy7cxnrx1f4v0gl53e182f",
-    "key": "fx2508ugpzk7urdi",
-    "size": 19451,
-    "response_time": 0.07393717765808105,
-    "metadata_service_time": 0.006797313690185547,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 14:38:33,552",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "pi96zmvmxdssw8mdv3p93rshu3eyaize",
-    "key": "ah4yoxcy5gn0uqwe",
-    "size": 19720,
-    "response_time": 0.04465198516845703,
-    "metadata_service_time": 0.007737874984741211,
-    "peer_id": "mictlanx-peer-0"
-}
-
-{
-    "timestamp": "2024-06-19 14:40:46,456",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "5r39ufs4kpyh29as33rdxzwe6a3c0or8",
-    "key": "ao7vdct8sei4gmrc",
-    "size": 19793,
-    "response_time": 0.06275415420532227,
-    "metadata_service_time": 0.007498979568481445,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 14:42:40,115",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "pb9b8lu5gasitffw2uhqtbw2r8xd7auf",
-    "key": "9c0fmhf7w9ira0ii",
-    "size": 19973,
-    "response_time": 0.044951677322387695,
-    "metadata_service_time": 0.008175134658813477,
-    "peer_id": "mictlanx-peer-0"
-}
-
-{
-    "timestamp": "2024-06-19 14:54:24,930",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "e6bye68q3t3c96p5dcos5jgqi1jx2miu",
-    "key": "rynbudw6ss0idqg5",
-    "size": 19864,
-    "response_time": 0.062108755111694336,
-    "metadata_service_time": 0.009041309356689453,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 15:08:06,175",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "gpux55svnze9zw76561r30ykgfdluoom",
-    "key": "ukc4glwa3hqwjgyh",
-    "size": 20091,
-    "response_time": 0.06317567825317383,
-    "metadata_service_time": 0.028500080108642578,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 15:17:40,329",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "m1fogrgovpg0sfeawgxbasrpxiu7oryy",
-    "key": "f7l2yzk2q5zv9iqq",
-    "size": 19982,
-    "response_time": 0.07182455062866211,
-    "metadata_service_time": 0.007383584976196289,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 15:25:58,384",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "s25lrat9ctrdl7j7qzuqrb71yxajf2i3",
-    "key": "udzodo4b0lmp3snp",
-    "size": 20568,
-    "response_time": 0.0611872673034668,
-    "metadata_service_time": 0.02728867530822754,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 15:33:27,557",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "4e9u19g6gq235ehr3fe2g4ur2cwdj6ge",
-    "key": "xm9ewzf2838hbtyp",
-    "size": 20385,
-    "response_time": 0.06009626388549805,
-    "metadata_service_time": 0.007134437561035156,
-    "peer_id": "mictlanx-peer-1"
-}
-
-{
-    "timestamp": "2024-06-19 15:35:07,511",
-    "level": "INFO",
-    "logger_name": "client-0",
-    "thread_name": "MainThread",
-    "event": "GET",
-    "bucket_id": "7dkdz9x0esc06tdql1kgdljck5trpxjg",
-    "key": "voet6u7hhpeu3raj",
-    "size": 20483,
-    "response_time": 0.061632633209228516,
-    "metadata_service_time": 0.027150392532348633,
-    "peer_id": "mictlanx-peer-1"
-}
-
